+++ rs/boundary_node/ic_boundary/src/salt_fetcher.rs
@@ -6,15 +6,13 @@ use std::{
 use anyhow::Error;
 use arc_swap::ArcSwapOption;
 use async_trait::async_trait;
-use candid::Principal;
-use candid::{Decode, Encode};
+use candid::{Decode, Encode, Principal};
+use ic_agent::Agent;
 use ic_bn_lib::prometheus::{
     register_int_counter_vec_with_registry, register_int_gauge_with_registry, IntCounterVec,
     IntGauge, Registry,
 };
 use ic_bn_lib::tasks::Run;
-use ic_canister_client::Agent;
-use ic_types::CanisterId;
 use salt_sharing_api::{GetSaltError, GetSaltResponse};
 use tokio::{
     select,
@@ -25,15 +23,6 @@ use tracing::warn;
 
 const SERVICE: &str = "AnonymizationSaltFetcher";
 const METRIC_PREFIX: &str = "anonymization_salt";
-
-fn nonce() -> Vec<u8> {
-    SystemTime::now()
-        .duration_since(SystemTime::UNIX_EPOCH)
-        .unwrap()
-        .as_nanos()
-        .to_le_bytes()
-        .to_vec()
-}
 
 struct Metrics {
     last_successful_fetch: IntGauge,
@@ -71,7 +60,7 @@ impl Metrics {
 
 pub struct AnonymizationSaltFetcher {
     agent: Agent,
-    canister_id: CanisterId,
+    canister_id: Principal,
     polling_interval: Duration,
     anonymization_salt: Arc<ArcSwapOption<Vec<u8>>>,
     metrics: Metrics,
@@ -87,7 +76,7 @@ impl AnonymizationSaltFetcher {
     ) -> Self {
         Self {
             agent,
-            canister_id: CanisterId::try_from_principal_id(canister_id.into()).unwrap(),
+            canister_id,
             anonymization_salt,
             polling_interval,
             metrics: Metrics::new(registry),
@@ -104,23 +93,17 @@ impl AnonymizationSaltFetcher {
 
         let query_response = match self
             .agent
-            .execute_update(
-                &self.canister_id,
-                &self.canister_id,
-                "get_salt",
-                Encode!().unwrap(),
-                nonce(),
-            )
+            .update(&self.canister_id, "get_salt")
+            .with_arg(Encode!().unwrap())
+            .call_and_wait()
             .await
         {
-            Ok(response) => match response {
-                Some(response) => response,
-                None => {
-                    update_fetch_metric("failure", "empty_response");
-                    warn!("{SERVICE}: got empty response from the canister");
-                    return;
-                }
-            },
+            Ok(response) if !response.is_empty() => response,
+            Ok(_) => {
+                update_fetch_metric("failure", "empty_response");
+                warn!("{SERVICE}: got empty response from the canister");
+                return;
+            }
             Err(err) => {
                 update_fetch_metric("failure", "update_call_failure");
                 warn!("{SERVICE}: failed to get salt from the canister: {err:#}");