+++ rs/registry/canister/src/invariants/routing_table.rs
@@ -2,10 +2,13 @@ use crate::invariants::common::{InvariantCheckError, RegistrySnapshot};
 
 use std::convert::TryFrom;
 
+use ic_base_types::CanisterId;
 use ic_protobuf::registry::routing_table::v1::{
     CanisterMigrations as pbCanisterMigrations, RoutingTable as pbRoutingTable,
 };
-use ic_registry_keys::{make_canister_migrations_record_key, make_routing_table_record_key};
+use ic_registry_keys::{
+    make_canister_migrations_record_key, make_canister_ranges_key, make_routing_table_record_key,
+};
 use ic_registry_routing_table::{CanisterMigrations, RoutingTable};
 use prost::Message;
 
@@ -19,14 +22,40 @@ pub(crate) fn check_routing_table_invariants(
 
 // Return routing table from snapshot
 fn get_routing_table(snapshot: &RegistrySnapshot) -> RoutingTable {
-    match snapshot.get(make_routing_table_record_key().as_bytes()) {
-        Some(routing_table_bytes) => {
-            let routing_table_proto =
-                pbRoutingTable::decode(routing_table_bytes.as_slice()).unwrap();
-            RoutingTable::try_from(routing_table_proto).unwrap()
-        }
-        None => panic!("No routing table in snapshot"),
+    // TODO(NNS1-3781): Remove this once we have sharded table supported by all clients.
+    let rt_from_routing_table_record =
+        match snapshot.get(make_routing_table_record_key().as_bytes()) {
+            Some(routing_table_bytes) => {
+                let routing_table_proto =
+                    pbRoutingTable::decode(routing_table_bytes.as_slice()).unwrap();
+                RoutingTable::try_from(routing_table_proto).unwrap()
+            }
+            None => panic!("No routing table in snapshot"),
+        };
+
+    // If there are shards, they should match the routing table record.
+    let shards = get_routing_table_shards(snapshot);
+    if !shards.is_empty() {
+        let rt_from_shards = RoutingTable::try_from(shards).unwrap();
+        assert_eq!(
+            rt_from_shards, rt_from_routing_table_record,
+            "Routing tables from shards and routing table record do not match."
+        );
     }
+
+    rt_from_routing_table_record
+}
+
+fn get_routing_table_shards(snapshot: &RegistrySnapshot) -> Vec<pbRoutingTable> {
+    let start = make_canister_ranges_key(CanisterId::from_u64(0)).into_bytes();
+    let end = make_canister_ranges_key(CanisterId::from_u64(u64::MAX)).into_bytes();
+    let mut shards = vec![];
+    for (_, value) in snapshot.range(start..=end) {
+        let routing_table_proto = pbRoutingTable::decode(value.as_slice()).unwrap();
+        shards.push(routing_table_proto);
+    }
+
+    shards
 }
 
 /// Iff `canister_migrations` is present, check that its invariants hold if reading and conversion succeed.
@@ -290,5 +319,36 @@ mod tests {
 
         assert!(check_routing_table_invariants(&snapshot).is_ok());
         assert!(check_canister_migrations_invariants(&snapshot).is_err());
+    }
+
+    #[test]
+    #[should_panic(expected = "Routing tables from shards and routing table record do not match.")]
+    fn if_sharded_ranges_they_must_match_original_routing_table() {
+        let mut snapshot = RegistrySnapshot::new();
+
+        // The routing table before canister migration.
+        let routing_table = RoutingTable::try_from(btreemap! {
+            CanisterIdRange{ start: CanisterId::from(0x0), end: CanisterId::from(0xff) } => subnet_test_id(1),
+            CanisterIdRange{ start: CanisterId::from(0x100), end: CanisterId::from(0x1ff) } => subnet_test_id(2),
+            CanisterIdRange{ start: CanisterId::from(0x200), end: CanisterId::from(0x2ff) } => subnet_test_id(3),
+         }).unwrap();
+
+        let routing_table = PbRoutingTable::from(routing_table);
+        let key1 = make_routing_table_record_key();
+        let value1 = routing_table.encode_to_vec();
+
+        let rt_shard = RoutingTable::try_from(btreemap! {
+            CanisterIdRange{ start: CanisterId::from(0x0), end: CanisterId::from(0xff) } => subnet_test_id(1),
+            CanisterIdRange{ start: CanisterId::from(0x100), end: CanisterId::from(0x1ff) } => subnet_test_id(2),
+        }).unwrap();
+
+        let rt_shard = PbRoutingTable::from(rt_shard);
+        let key2 = make_canister_ranges_key(CanisterId::from_u64(0x0));
+        let value2 = rt_shard.encode_to_vec();
+
+        snapshot.insert(key1.into_bytes(), value1);
+        snapshot.insert(key2.into_bytes(), value2);
+
+        check_routing_table_invariants(&snapshot).unwrap();
     }
 }