+++ rs/tests/consensus/subnet_recovery/sr_app_failover_nodes_test.rs
@@ -1,7 +1,7 @@
 use anyhow::Result;
 
 use ic_consensus_system_test_subnet_recovery_common::{
-    setup_failover_nodes as setup, test_without_tecdsa as test,
+    setup_failover_nodes as setup, test_without_chain_keys as test,
 };
 use ic_system_test_driver::driver::group::SystemTestGroup;
 use ic_system_test_driver::systest;