+++ rs/nns/cmc/CHANGELOG.md
@@ -11,6 +11,15 @@ here were moved from the adjacent `unreleased_changelog.md` file.
 INSERT NEW RELEASES HERE
 
 
+# 2025-05-31: Proposal 136797
+
+http://dashboard.internetcomputer.org/proposal/136797
+
+## Changed
+
+- Use the mint_cycles128 system API, so larger amounts of cycles can now be minted.
+
+
 # 2025-02-07: Proposal 135205
 
 http://dashboard.internetcomputer.org/proposal/135205